Quick refactor, broke out the menu functionality into its own file since it's been growing quite a bit
