+++ software/display/display.h
@@ -183,58 +183,6 @@ typedef struct inputState {
 } inputState_t;
 
 /**
- * @brief Type to handle the menu options
- */
-typedef struct menuOption {
-    /**
-     * @param pOptionString Pointer to the string for this option.
-     * This is a required option.
-     */
-    char *pOptionString;
-    /**
-     * @param pSubMenu Pointer to a sub-menu
-     * If this is NULL, then no sub-menu exists.
-     */
-    void *pSubMenu;
-    /**
-     * @param pUpdateFun Pointer to a function to update this item
-     * This is the function that is called when this menu option
-     * is selected.
-     * If NULL, then no update function exists, and it has to be a sub-menu.
-     */
-    void *pUpdateFun;
-    /**
-     * @param pDisplayFun Pointer to a function that returns the string
-     * which shows the current option for this function that is selected.
-     * If NULL, then no update function exists, and it has to be a sub-menu.
-     * This function takes in a string as an argument, to return to the display
-     * function.
-     */
-    void *pDisplayFun;
-
-} menuOption_t;
-
-/**
- * @brief Struct holding the menu state.
- */
-typedef struct menuState {
-    /**
-     * @param pUpperMenu Pointer to the option list one step above. This is to
-     * get back. NULL if this is currently the top level menu. This is a
-     * menuState_t pointer.
-     */
-    void *pUpperMenu;
-    /**
-     * @param pMenuOptionList Pointer to the menu list
-     */
-    const menuOption_t *pMenuOptionList;
-    /**
-     * @param pCurrentMenuOption Pointer to the current option
-     */
-    menuOption_t *pCurrentMenuOption;
-} menuState_t;
-
-/**
  * @brief Struct holding the display state, along with calculator
  * results and user input states which should be displayed.
  * This script and the underlying structs are populated
@@ -280,7 +228,7 @@ typedef struct displayState {
     /**
      * @param pMenuState Pointer to menu state.
      */
-    menuState_t *pMenuState;
+    void *pMenuState;
     /**
      * @param printedInputBuffer Input buffer printed by calc core.
      */
@@ -294,10 +242,6 @@ typedef struct displayState {
 extern displayState_t displayState;
 extern xSemaphoreHandle displayStateSemaphore;
 extern EventGroupHandle_t displayTriggerEvent;
-
-// Menu related variables
-extern menuState_t bitSizesMenu;
-extern menuState_t topMenu;
 
 /* -------------------------------------------
  * ---------- FUNCTION PROTOTYPES ------------
@@ -317,16 +261,5 @@ void initDisplayState(displayState_t *pDisplayState);
 void displayTask(void *p);
 
 void testDisplay(void);
-/**
- * @brief Function to get the current font.
- * @param p Pointer to the task argument. Not used for now
- * @return Nothing
- */
-void getCurrentFont(char *pString);
-/**
- * @brief Function to iterate through the current fonts.
- * @param p Pointer to the task argument. Not used for now
- * @return Nothing
- */
-void changeFont(void);
+
 #endif /* DISPLAY_H_ */
